+++ src/runtimes/westend.rs
@@ -41,7 +41,7 @@ use std::{
 use subxt::{
     sp_core::{sr25519, Pair as PairT},
     sp_runtime::AccountId32,
-    DefaultConfig, PolkadotExtrinsicParams, PairSigner,
+    DefaultConfig, PairSigner, PolkadotExtrinsicParams,
 };
 
 #[subxt::subxt(
@@ -56,7 +56,8 @@ use node_runtime::{
     utility::events::BatchInterrupted, utility::events::ItemCompleted,
 };
 
-pub type Api = node_runtime::RuntimeApi<DefaultConfig, PolkadotExtrinsicParams<DefaultConfig>>;
+pub type Api =
+    node_runtime::RuntimeApi<DefaultConfig, PolkadotExtrinsicParams<DefaultConfig>>;
 
 type Call = node_runtime::runtime_types::westend_runtime::Call;
 type StakingCall = node_runtime::runtime_types::pallet_staking::pallet::pallet::Call;
@@ -149,10 +150,7 @@ pub async fn try_run_batch(
     let seed = fs::read_to_string(config.seed_path)
         .expect("Something went wrong reading the seed file");
     let seed_account: sr25519::Pair = get_from_seed(&seed, None);
-    let seed_account_signer =
-        PairSigner::<DefaultConfig, PolkadotExtrinsicParams<DefaultConfig>, sr25519::Pair>::new(
-            seed_account.clone(),
-        );
+    let seed_account_signer = PairSigner::<DefaultConfig, sr25519::Pair>::new(seed_account.clone());
     let seed_account_id: AccountId32 = seed_account.public().into();
 
     // Get signer account identity
@@ -258,11 +256,12 @@ pub async fn try_run_batch(
                 let batch_response = api
                     .tx()
                     .utility()
-                    .batch(calls_for_batch_clipped.clone())
-                    .sign_and_submit_then_watch(&seed_account_signer)
+                    .batch(calls_for_batch_clipped.clone())?
+                    .sign_and_submit_then_watch_default(&seed_account_signer)
                     .await?
                     .wait_for_finalized()
                     .await?;
+
                 debug!("batch_response {:?}", batch_response);
 
                 // Alternately, we could just `fetch_events`, which grabs all of the events like